+++ include/layers/layer_data.hpp
@@ -24,7 +24,7 @@ namespace NeuralNet
         void resize(size_t train_num, size_t data_num);
 
         /* returns the desired array */
-        double *get(DataIndex idx) const;
+        float *get(DataIndex idx) const;
 
         /* returns the dimensions */
         size_t getDataNum() const { return m_data_num; }
@@ -32,7 +32,7 @@ namespace NeuralNet
 
     private:
         size_t m_train_num, m_data_num;
-        double *data;
+        float *data;
         static constexpr size_t DATA_COUNT = static_cast<int>(DataIndex::END)
                 - static_cast<int>(DataIndex::START) + 1;
     };
