+++ include/layers/max_pool_layer.hpp
@@ -37,8 +37,8 @@ namespace NeuralNet
         virtual std::string what() { return "maxpool"; }
         virtual size_t getNeuronNum() const;
 
-        virtual void setLearnRate(double rate) {}
-        virtual double getLearnRate() const { return 0; }
+        virtual void setLearnRate(float rate) {}
+        virtual float getLearnRate() const { return 0; }
 
     private:
         const Dimension m_dim;
